network-testbed/requirements.txt
# Network Testbed – Python dependencies
pyyaml>=6.0
docker>=7.0.0
pytest>=8.0.0
pytest-html>=4.1.0
pytest-timeout>=2.2.0
jinja2>=3.1.0
requests>=2.31.0
